// components/console/src/pages/Topology/Topology.enum.ts
export enum RoutesPaths {
  Topology = '/topology'
}

export enum QueriesTopology {
  GetTopologyData = 'get-topology-data-query',
  GetNetworkOverview = 'get-network-overview-query',
  GetBackbones = 'get-backbones-topology-query',
  GetSites = 'get-sites-topology-query',
  GetVans = 'get-vans-topology-query'
}