// components/console/src/pages/Vans/Vans.enum.ts
export enum RoutesPaths {
  Vans = '/Vans'
}

export enum QueriesVans {
  GetVans = 'get-vans-query',
  GetVan = 'get-van-query',
  GetInvitations = 'get-invitations-query',
  GetInvitationsYAML = 'get-invitations-yaml-query',
  GetMembers = 'get-members-query',
  GetVanMembers = 'get-van-members-query',
  GetVanInvitations = 'get-van-invitations-query',
  GetAccessClaim = 'get-access-claim-query',
  GetAccessMember = 'get-access-member-query'
}