// components/console/src/pages/Libraries/Libraries.enum.ts
export enum RoutesPaths {
  Libraries = '/libraries'
}

export enum QueriesLibraries {
  GetLibraries = 'GetLibraries',
  GetLibrary = 'GetLibrary',
  GetLibraryBlocks = 'GetLibraryBlocks',
  GetLibraryBlock = 'GetLibraryBlock',
  GetLibraryBody = 'GetLibraryBody',
  GetLibraryHistory = 'GetLibraryHistory',
  GetBlockTypes = 'GetBlockTypes',
  GetBodyStyles = 'GetBodyStyles'
}